BACKEND/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
PyMySQL==1.1.0
mysqlclient==2.2.0
PyJWT==2.10.1
django-filter==23.4
django-extensions==3.2.3
Pillow==10.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
black==23.12.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
drf-spectacular==0.26.5
